+++ src/app/profile/page.tsx
@@ -1,4 +1,4 @@
-import ProfilePage from '@/features/profile/components/ProfilePage';
+import ProfilePage from '@/features/profile/container/ProfilePage';
 import React from 'react';
 
 const Profile = () => {
