+++ src/features/profile/components/exchange/index.ts
@@ -1,0 +1,4 @@
+export { default as ExchangeContents } from './ExchangeContents';
+export { default as MyRegisteredBooks } from './MyRegisteredBooks';
+export { default as TradeRecords } from './TradeRecords';
+export { default as TradeReviews } from './TradeReviews';
