+++ src/features/profile/components/profile/index.ts
@@ -1,0 +1,2 @@
+export { default as ProfileEditModal } from './ProfileEditModal';
+export { default as Profile } from './Profile';
